chore: rename files to camelcase
