+++ frontend/src/Editor/LeftSidebar/index.js
@@ -2,10 +2,10 @@ import '@/_styles/left-sidebar.scss';
 
 import React, { useState } from 'react';
 
-import { LeftSidebarItem } from './sidebar-item';
-import { LeftSidebarInspector } from './sidebar-inspector';
-import { LeftSidebarDataSources } from './sidebar-datasources';
-import { LeftSidebarZoom } from './sidebar-zoom';
+import { LeftSidebarItem } from './SidebarItem';
+import { LeftSidebarInspector } from './SidebarInspector';
+import { LeftSidebarDataSources } from './SidebarDatasources';
+import { LeftSidebarZoom } from './SidebarZoom';
 import { DarkModeToggle } from '../../_components/DarkModeToggle';
 import useRouter from '../../_hooks/use-router';
 import { LeftSidebarDebugger } from './SidebarDebugger';
